requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
requests==2.26.0
#rust based linter
ruff==0.0.284
boto3==1.24.87

#web
fastapi==0.85.0
uvicorn==0.18.3

#databricks
pandas
numpy
requests
databricks-sql-connector
databricks
pyspark
pytest
python-dotenv
matplotlib